[TASK] Use RouteResult in EXT:reactions resolver
The new Backend Routing feature
with dynamic placeholders was added
just before EXT:reactions was
introduced. This way, the Reactions
Resolver class (responsible for
validating user input for reactions,
and checking for the reactions identifier)
is now using the RouteResult object,
in order to get rid of manual URL
resolving logic.

Resolves: #99262
Releases: main
Change-Id: Ife744069cb788dfdb1113ec8f68d47e7a084fbdd
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/76916
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Tested-by: Oliver Bartsch <bo@cedev.de>