[BUGFIX] Handle reaction records with missing reaction class
When reaction records are present in a system and their
corresponding implementation class is not, the reaction
module failed due to fluid parser errors.

This is now handled properly. A badge with corresponding
information is displayed for such records and functions
like "edit record" are disabled, since they would lead to
further errors.

Additionally, this also fixes the hide / unhide control,
which previously did not work due to usage of a non-
existent magic getter.

Resolves: #99322
Related: #98373
Releases: main
Change-Id: I8ccb9ccffeedf4b5f3880f06d968880c4d5232f8
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/77000
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>